git-svn-id: file:///mnt/data/Workspace/zie87/game_development/tmp/hare-game-engine/p/trunk@315 67360d68-41fe-4630-88ca-a5c15c9c3197

+++ graphics/ComponentSprite.cpp
@@ -42,8 +42,8 @@ namespace hare
     {
         Sprite::List::iterator it = components.begin();
         for (;it != components.end(); ++it){
-            int origoX = origoPos.x;
-            int origoY = origoPos.y;
+            float origoX = origoPos.x;
+            float origoY = origoPos.y;
 
             if (!bFaceX){
                 origoX = -origoX;
